Remove bug in grep pattern
The pid the `grep -v` is trying to eliminate is not at the end
of the line.

`all_tmux_processes` produces the following output (for example):

```
tmux                        32069 29270
```

Here the first number is the pid and the second -- the uid. So, since
the idea of `number_tmux_processes_except_current_server` is to eliminate
32069 from consideration, it should look for "<SPACE>PID<SPACE>".