Fix systemd unit install on clean system
If the .config/systemd/user directory did not exist, it would not get
created, as the mkdir command was inside the function
systemd_unit_file() which is piped into the target file. The shell
determined that file could not be created first (as the directory did
not exist), so never called the function.

Signed-off-by: Paul Elliott <paul@niburu.org>